move and rename param_validation for clarity
